+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
  - gitignore
+ - ft_rbtnew [rbtree]
 
 ### Changed
 ### Fixed